added clap for pc cli
